ch4/workq: uninlining workq progress
Inlining workq progress will cause inlining too deep, severely slow down
compilation while the overhead of the single function call is relatively
cheap.
