+++ src/mpid/ch4/src/Makefile.mk
@@ -22,6 +22,8 @@ noinst_HEADERS += src/mpid/ch4/src/ch4_comm.h     \
                   src/mpid/ch4/src/ch4r_init.h    \
                   src/mpid/ch4/src/ch4r_proc.h    \
                   src/mpid/ch4/src/ch4i_comm.h    \
+                  src/mpid/ch4/src/ch4i_workq.h   \
+                  src/mpid/ch4/src/ch4i_workq_types.h \
                   src/mpid/ch4/src/ch4r_recvq.h   \
                   src/mpid/ch4/src/ch4r_recv.h    \
                   src/mpid/ch4/src/ch4r_callbacks.h     \
@@ -37,6 +39,7 @@ mpi_core_sources += src/mpid/ch4/src/ch4_globals.c        \
                     src/mpid/ch4/src/ch4_progress.c       \
                     src/mpid/ch4/src/ch4_win.c            \
                     src/mpid/ch4/src/ch4i_comm.c          \
+                    src/mpid/ch4/src/ch4i_workq_progress.c \
                     src/mpid/ch4/src/ch4r_init.c          \
                     src/mpid/ch4/src/ch4r_proc.c          \
                     src/mpid/ch4/src/ch4r_recvq.c         \
